requirements.txt
discord.py
requests
PyMySQL